+++ pkg/compiler/lib/src/commandline_options.dart
@@ -68,6 +68,8 @@ class Flags {
   static const String readData = '--read-data';
   static const String writeData = '--write-data';
 
+  static const String serverMode = '--server-mode';
+
   /// Flag for a combination of flags for 'production' mode.
   static const String benchmarkingProduction = '--benchmarking-production';