+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.2.0-dev.XX.0
+
+### Dart for the Web
+
+#### dart2js
+
+* The `--categories=*` flag is being replaced. `--categories=all` was only used
+  for testing and it is no longer supported. `--categories=Server` continues to
+  work at this time but it is deprecated, please use `--server-mode` instead.
+
 ## 2.1.0 - 2018-10-31
 
 This is a minor version release. The team's focus was mostly on improving